Make url-testing nicer
Basically, we only want to tell the user when there's an _external_
site redirect. So, from the same root domain (e.g., "xyz.google.com"
to "abc.google.com"), we shouldn't say anything. This should cause
the bot to be a bit less spammy.

Also, renamed "original_url" and "transformed" on SlackURL to "text"
and "url", which is a bit clearer. One is the display text, the other
is the link behind the text, disregarding the process of how it got
there.